x/hub: Remove configurable eviction policy.
This feature was too complicated for its value.

Even for the common case of evicting the oldest, it copies the whole queue for
each eviction, which is common case for a slow reader, exactly the case we're
trying to reduce the cost of. Making a special case of that would make it even
more complicated.

Also, calling back into user code allows it to block the hub. We could just
document that isn't allowed, but Hyrum is right over there. I wouldn't drop
this feature just for that, but this was another nail.